[SPARK-58941][SDP] Sort flows by identifier parts, and move the test into SchemaInferenceUtilsSuite
Address review feedback on the schema-inference flow sort:

- Sort on the identifier's parts `(catalog, database, table)` instead of `quotedString`.
  Both are injective (no dotted-name collision), but the parts tuple also preserves the
  existing order for dotless identifiers, whereas `quotedString` reordered some pairs
  (back-tick 0x60 sorts after name characters, dot 0x2E before them). Flow identifiers are
  always fully qualified, so all three parts are present and the ordering never depends on
  an absent catalog/database.
- Move the `inferSchemaFromFlows` test into the existing `SchemaInferenceUtilsSuite`
  (now `QueryTest with SharedSparkSession`) and delete the single-test suite; reword it
  around dotted-name folding and trim the comments.
